added an ExternalMethod to disable the old content types and rename the new ones to the old names. its less confusing for the common user.
this external methods are registered in portal on install and removed on
uninstall.  for future a configlet would be nice to start and stop this
methods would be a nice feature.

also someone should please check the security of the whole. iatm i'am not
very familiar with security on ExternalMethods.


git-svn-id: https://svn.plone.org/svn/collective/trunk@36 db7f04ef-aaf3-0310-a811-c281ed44c4ad